[tests] Do not check the MaxRequestContentBufferSize anymore
In the AndroidClientHandlerTests.cs are based on mono's
HttpClientHandlerTest [1]. One of the things to test is
MaxRequestContentBufferSize. The code for this property changed
recently in mono [2]. So we will do the same here.

It is not clear why this returns 0. The [3] states, that returning 0
means no limit. The documentation doesn't talk about that though. [4]

[1] https://github.com/xamarin/monodroid/commit/b9ca62997206b89101344bad76b0dc0be899cc91
[2] https://github.com/mono/mono/commit/2f411750acfd2132f9b4970ef8cf11d9a0f62df9#diff-64cf37c8688f8a48476999ba2dd52ab7R184
[3] https://github.com/mono/mono/commit/2f411750acfd2132f9b4970ef8cf11d9a0f62df9#diff-a94223e905e684e2f789149da44fafd9R16
[4] https://docs.microsoft.com/en-us/dotnet/api/system.net.http.httpclienthandler.maxrequestcontentbuffersize?view=netframework-4.8