[xaprepare] Use NuGet binary we download
`NuGetRunner` doesn't properly specify full path to the nuget binary we download
as part of the prepare run and, thus, it depends on any random version of the
utility in the user's `$PATH` (if at all) which may lead to random errors
depending on the nuget version. Fix the issue by passing the full path to the
downloaded binary.

Additionally, fix `ToolRunner` to properly add arguments when managed
runtime (`mono`) is used on Unix. Until now any initial arguments were passed
to **mono** in such case, instead of to the managed executable. Doh.