[xaprepare] Fix bundle name generation for PRs
Context: https://github.com/microsoft/azure-pipelines-agent/issues/2330

We've been running into an issue on Azure Pipelines where the bundle
created in our 'Prepare' stage has a different name than the one that is
calculated later during build stages. This is because we're building a
merge commit on Azure pipelines, and if the target branch of the PR
changes in between stages, the merge commit that is checked out will
vary. We'll attempt to work around this (hopefully temporarily) by using
the HEAD of the source branch of a PR to calculate the bundle hashes,
rather than the HEAD of the source that is checked out at any given time.