[tests] fix for ExistingProject (#3292)
Context: https://github.com/xamarin/monodroid/pull/1007

We have been getting the following error on commercial builds:

	…/xamarin-android/bin/Release/bin/xabuild …/xamarin-android/bin/TestRelease/temp/HelloFormsFalse/HelloForms.sln /t:Restore /noconsolelogger "/flp1:LogFile=…/xamarin-android/bin/TestRelease/temp/HelloFormsFalse/build.log;Encoding=UTF-8;Verbosity=diagnostic" @"…/xamarin-android/bin/TestRelease/temp/HelloFormsFalse/project.rsp"
	Microsoft (R) Build Engine version 16.0.42-preview+g804bde742b for Mono
	Copyright (C) Microsoft Corporation. All rights reserved.

	MSBUILD : error MSB1009: Project file does not exist.
	Switch: …/xamarin-android/bin/TestRelease/temp/HelloFormsFalse/HelloForms.sln

In c00095b, the tests were passing on xamarin-android builds, but
they fail when running in monodroid builds?  This appears to be due
to differences in test ordering.

There was a code path where `ProjectBuilder` was deleting its
`ProjectDirectory` and calling `Populate()` on the project.

For an `ExistingProject` in `DeleteBinObjTest`, this would nuke the
directory we just extracted from a zip file--causing MSBuild to fail.

The fix is to add a new `ShouldPopulate` flag we can set to `false`
to skip the directory deletion.