[Xamarin.Android.Build.Tasks] Fix DesignTimeBuild & Resource changes (#3290)
One of the problems we have with the `ManagedResourceParser` is that
it uses the `R.txt` to generate the `Resource.Designer.cs` file.
The issue there is that if the user modifies a layout file (say adds
or changes an id) that will NOT be picked up until the next build.
This is because the `R.txt` file is generated by `aapt`/`aapt2` and
those tools are not called for DesignTime Builds.

What we need to do is pass in the `$(_AndroidResFlagFile)` into the
`<GenerateResourceDesigner/>` task.  This will allow us to do a last
modified date time check against the `R.txt` file.  If
`$(_AndroidResFlagFile)` is newer then we should NOT use `R.txt` and
should instead parse the layout files to pick up changes.