[xaprepare] Don't throw when console cursor coords are incorrect (#3312)
`System.Console` will throw an exception when the cursor coordinates
are out of range, but we don't want to check whether left/top are
less than the buffer width/height, respectively, because that might
throw as well, so we just catch the possible exception.

On Unix it may happen when the windows were just being re-created
(e.g. when switching between monitors) and there's no reason to stop
the app because of this so just log the exception and move on.