[build] Improve toolchain caching on Azure Pipelines
The Azure Pipelines hosted macOS machines we use for build and test jobs
already have Android SDK components installed. We should build and test
against that known location to improve our ability to cache Android
toolchain components, as most of what we need is already installed and
will be less prone to being cleaned up over time. This should reduce the
time cost of our dependency provisioning steps.

We're using a different agent pool type on the Windows side, so we should
continue to use the same toolchain location there.