[Xamarin.Andorid.Build.Tasks] Check if LogicalName is valid before check Resource File.
Some of our customers are generating `Resource` files
which have invalid file names (as far as aapt/aapt2 is
concerned). They use the `LogicalName` meta data to
provide the "Real" name that the build process should
use.

However when using `aapt2` our `CheckForInvalidResourceFileNames`
task was NOT using the `LogicalName` metadata. It was
just using the filename. This breaks their projects with
an

	APT0000: Invalid file name: It must contain only [^a-zA-Z0-9_.-]+

or

	APT0000: Invalid file name: It must contain only [^a-zA-Z0-9_.]+

We should really support this scenario since we allow the
use of `LogicalName` for resource files anyway.

This commit fixes that issue and adds a unit test.