Adjust wording to match the new order
Now that the steps for attaching GDB in Visual Studio come before the
steps for attaching LLDB in Android Studio, reword the overview
sentences for those sections to match that new order.

Now that mono/lldb-binaries is the first recommended option, add an
additional hint about how to attach to an app that is already running
under the managed debugger.