[docs] Add template file for draft release notes
The idea for this new `Documentation/release-notes/master.md` file is
that each time a new user-facing bug fix or feature is merged to
`master`, the author(s) of the change will add a corresponding draft
release note to this file.  The note can be added as part of the source
code pull request itself, or it can be added in a separate follow-up
pull request shortly afterward.

If the note for a new feature uses images, the images can be added to
`Documentation/release-notes/images/`.

Authors of new fixes and improvements don't need to do anything special
if their changes get cherry-picked from `master` to a release branch.
The team members who handle final release notes preparation and
publication will copy or move notes as needed from `master.md` to
earlier releases for changes that get cherry-picked.

The `master.md` template follows the same section order as the [final
published release notes][0], but it omits a few details like release
dates that aren't needed in the draft format.

The final notes also follow the text formatting rules described in the
[xamarin-docs contribution guidelines][1].  The draft notes don't need
to follow those guidelines too strictly.  For example, it's fine to use
backticks around paths in these draft notes rather than italics.  That
makes it easier to avoid unintentional formatting issues when paths
include characters like `_` or `*`.

Similarly, there's no need to worry about list indentation or line
wrapping in these draft notes.  Authors can optionally wrap lines to 80
characters and indent list items to match the example items, but it's
not required.

As part of branching for a new release, the current `master.md` file
will be renamed to match the new release branch name, and a new empty
template `master.md` file will be added by copying the original template
from this commit and updating the version numbers.

[0]: https://docs.microsoft.com/xamarin/android/release-notes/
[1]: https://github.com/MicrosoftDocs/xamarin-docs/blob/live/contributing-guidelines/template.md