[Xamarin.Android.Build.Tasks] Apk/ABI fails with TFN=v7.1 & Aapt2 (#3242)
Fixes: https://github.com/xamarin/xamarin-android/issues/3083

`aapt2` can be inconsistent regarding its exit code and writing
errors to stderr.  It turns out that there are cases where only
WARNINGS are written to stderr.  In this case the exit code is 0.
However we can not rely on the exit code only, since there have been
cases where the exit code is 0 but we have errors on stderr.

So what we need to do is parse the stderr output and look for those
`tags` which we know represent warnings.  This is so we can correctly
classify those errors/warnings to make sure we correctly report issues.