[Xamarin.Android.Build.Tasks] Fix issues with re running CilStrip
Fixes https://github.com/xamarin/xamarin-android/issues/1240

The `CilStrp` task was not able to handle being called
incrementally. When called a second time the resulting
error would be raised.

	error XA3001: System.IO.IOException: The file '/Users/dean/Projects/Foo5F/Foo/obj/Release/android/assets/non-stripped/Foo.dll' already exists.
		at System.IO.FileSystem.LinkOrCopyFile (System.String sourceFullPath, System.String destFullPath) [0x000b9] in <06589f46f7f64859943a6a461f88c56e>:0
		at System.IO.FileSystem.MoveFile (System.String sourceFullPath, System.String destFullPath) [0x0003a] in <06589f46f7f64859943a6a461f88c56e>:0
		at System.IO.File.Move (System.String sourceFileName, System.String destFileName) [0x00083] in <06589f46f7f64859943a6a461f88c56e>:0
		at Xamarin.Android.Tasks.CilStrip.DoExecute () [0x000a8] in <fd1a3c6c66de470d989dd70ca1f91267>:0
		at Xamarin.Android.Tasks.CilStrip.Execute () [0x00000] in <fd1a3c6c66de470d989dd70ca1f91267>:0

This commit make sure we do not try to strip an assembly
if it has already been stripped. It also adds a new
Unit Test to make sure the build system behaves as we expect.