[Xamarin.Android.Build.Tasks] <FilterAssemblies/> should look for .jar files
Fixes: https://github.com/xamarin/xamarin-android/issues/3343
Context: https://github.com/xamarin/XamarinComponents/blob/54646f5c0bef9c618c700628bd1167768d7c9b4f/Android/Guava/source/Guava.ListenableFuture/Guava.ListenableFuture.csproj

We introduced a performance improvement in 5ec3e3ab that "classifies"
assemblies as Xamarin.Android assemblies or not. This allows us to
skip processing of NetStandard/BCL assemblies throughout the build.

Unfortunately, this Guava.ListenableFuture assembly is odd:

* No `[assembly: TargetFramework]` attribute at all
* No reference to `Mono.Android.dll`

The only "marker" is that it has a `.jar` file as an
`EmbeddedResource`, which is put there by `<EmbeddedJar/>`. Its
weirdness is related to it not having any C# code, and using SDK-style
projects.

The fix here is to update `<FilterAssemblies/>` to look for:

* `[assembly: TargetFramework("MonoAndroid,Version=v9.0")]`
* `Mono.Android.dll` reference
* `EmbeddedResource` ending with `*.jar`
* `EmbeddedResource` beginning with `__Android`

I also reworked this task so that the logic is hardcoded. It does not
seem useful to pass in input parameters from MSBuild targets anymore.