[Xamarin.Android.Build.Tasks] fix Inputs for aapt2 (#3355)
Fixes: https://github.com/xamarin/xamarin-android/issues/3336

@brendanzagaeski found a really weird bug:

1. Build a project with aapt2 with an invalid XML file.
2. Remove the file from the project, build again.
3. Add the file back, build a third time.

On step 3 & beyond, no build error is given--meaning *something* must
be out of date.

I was able to reproduce the problem in a test, and saw it only occurs
with aapt2.

The `_CompileResources` MSBuild target was skipped:

    Skipping target "_CompileResources" because all output files are up-to-date with respect to the input files.
    Input files: Resources\values\ids.xml;Resources\drawable-hdpi\Icon.png;Resources\drawable-mdpi\Icon.png;Resources\drawable-xhdpi\Icon.png;Resources\drawable-xxhdpi\Icon.png;Resources\drawable-xxxhdpi\Icon.png;Resources\layout\Main.axml;Resources\values\Strings.xml
    Output files: obj\Debug\flata\\_CompileResources.stamp

But `_UpdateAndroidResgen` was *not* skipped:

    Building target "_UpdateAndroidResgen" completely.
    Input file "C:\src\xamarin-android\bin\TestDebug\temp\InvalidAndroidResourceTrue\obj\Debug\res\values\ids.xml" is newer than output file "obj\Debug\R.cs.flag".

Reviewing the `Inputs` of `_CompileResources`, it seems like we need
to include the following files to better match `_UpdateAndroidResgen`:

* `$(MSBuildAllProjects)` - so if the `.csproj` changes.
* `$(_AndroidBuildPropertiesCache)` - so if any of the important
  Xamarin.Android MSBuild properties change.

After adding these `Inputs`, I also had to add a `Condition` to skip
the `_CompileResources` target entirely if `@(AndroidResource)` was
blank. Otherwise projects like `Xamarin.Android.NUnitLite` did not
build. This was skipping before because the `Inputs` were empty.