[Build] Remove libzip and LibZipSharp
LibZipSharp is now available as a nuget package.
This commit removes the submodules for `libzip`
and `LibZipSharp` in favour of using the nuget
package.

The `LibZipSharp` package contains the native
binaries for `libzip` however by default it
does NOT copy these to the output directory.
In order for that to happen the following MSBuild
property needs to be set

	<LibZipSharpBundleAllNativeLibraries>true</LibZipSharpBundleAllNativeLibraries>

This means the native binaries will be copied to the
`$(OutputPath)\<PLATFORM>` directories where `PLATFORM`
will be `Darwin`, `Linux` or `Windows`.

This means we need to update our build system to make sure
we include the files from the correct locations. We also
need to make sure we call `Restore` on projects using
`LibZipSharp` now as the package will need to be restored.

Another change is that we have removed the versioning from the
`libzip` native libraries. They used to be `libzip.5.0.dylib`
but are now just `libzip.dylib`. This was done upstream in
the `LibZipSharp` package itself to allow for support for
.Net Core based `DllImport` which do NOT use a `dll.config`
file, but instead use only the dll name (e.g `libzip`).

We are also including the `Linux` version of the binaries
in the installer.