[Xamarin.Android.Build.Tasks] fix the NuGetizer 3000
Fixes: http://work.azdo.io/917139
Context: https://github.com/NuGet/Home/wiki/NuGetizer-Core-Features

The NuGetizer 3000, was failing on Xamarin.Android 9.4 with:

    The "FilterAssemblies" task was not given a value for the required parameter "DesignTimeBuild".

This tool runs a `GetPackageContents` MSBuild target that is defined
here:

    https://github.com/NuGet/NuGet.Build.Packaging/blob/dev/src/Build/NuGet.Build.Packaging.Tasks/NuGet.Build.Packaging.targets#L201

Since this isn't a `Build` or a `Compile`, neither of these targets
will run:

* `_SetupDesignTimeBuildForBuild`
* `_SetupDesignTimeBuildForCompile`

I don't think we can make the `_AddAndroidCustomMetaData` MSBuild
target (that runs `<FilterAssemblies/>`) depend on either of these.

The simple fix is to just remove `[Required]` from the
`DesignTimeBuild` property, as it will be fine for it to default to
`false`.