[tests] Change parallel scope for some multi source tests
We're hitting NuGet restore related issues on macOS when running tests
with multiple test case sources in parallel. I believe that this is not
occurring on Jenkins because all NuGet depdencies for tests there have
been cached for quite some time. On Azure Pipelines, we're more often
dealing with clean machines, and parallel attempts to restore/install
into global cache paths are occasionally failing:

    NuGet.targets(121,5): error : Could not find file "/Users/vsts/.local/share/NuGet/v3-cache/1ca707a4d90792ce8e42453d4e350886a0fdaa4d$ps:_api.nuget.org_v3_index.json/nupkg_sqlitepclraw.core.1.1.8.dat"
    NuGet.targets(121,5): error : The file '/Users/vsts/.nuget/packages/xamarin.android.support.compat/28.0.0.1/.nupkg.metadata' already exists.
    NuGet.targets(121,5): error : Could not find file "/Users/vsts/.nuget/packages/xamarin.forms/3.4.0.1008975/2omeck77.h7w"

The fix is to update the `ParallelScope` attribute associated with these
affected tests to ensure that dynamically generated instances don't run
in parallel with eachother. This still allows for parallel execution
against other tests or fixtures.