[Xamarin.Android.Build.Tasks] Parse R.txt int styleable IDs as base-ten
Fixes: https://github.com/xamarin/xamarin-android/issues/3395

For some reason, `aapt` and `aapt2` generate base-ten numbers instead of
hexadecimal numbers for the `int styleable` IDs in the `R.txt` file.
All the other IDs are hexadecimal.

Correct the managed resource parser to read the `int styleable` IDs as
base-ten numbers instead of base-sixteen numbers.  Adjust the generated
comment text for `int styleable` IDs accordingly too.

Add a test for this scenario, and update the expected outputs for the
existing tests to account for the adjusted comment text.