Added _AddAndroidDefines target dependency for _ManagedUpdateAndroidResgen
When running deferred builds as following:

`/t:DeferredBuild /p:DeferredBuild=true /p:BuildingInsideVisualStudio=true`

`ManagedDesignTimeBuild` property will be set to `true` and `_ManagedUpdateAndroidResgen`
target will be invoked as part of the `UpdateAndroidResources` target.

At this time we need to make sure that `_AddAndroidDefines` target is executed before
creating the addtional resource cache in order to set the required `AndroidNDKDirectory`
property.