[generator] Use a lookup cache in SymbolTable to find types (#452)
The way we use a `SymbolTable` is to fully populate it with types and
then perform lots of lookups on it.  We can optimize this by building
a lookup cache when we start accessing the types and then using it
for all subsequent lookups.

If any additional types are added to the table, we clear the cache so
it will be rebuilt.  In practice this doesn't happen often, so we are
not constantly rebuilding the cache.  (In testing the most we build
the cache is twice, out of thousands of accesses.)

Timing:

  * `class-parse` invocations:
      * commit 6373f517:    5563ms
      * With caching:       2555ms

  * `generator --only-xml-adjuster` invocations:
      * commit 6373f517:    2645ms
      * With caching:        782ms