Bump $(ProductVersion) to 10.0.99
We (try to) base Xamarin.Android product versions on the latest
Android versions.  The system images for API-29 mention "Android 10",
so with the merging of stable API-29 support (936a09d1) we want our
first release with API-29 support to likewise be v10.0.

The [d16-3][0] branch is the first release which will offer stable
API-29 support, and was bumped to report itself as [v10.0][1].

As xamarin-android/master is tracking the *next* Xamarin.Android
version, bump `$(ProductVersion)` to v10.0.99 for eventual release
as v10.1.0.

[0]: https://github.com/xamarin/xamarin-android/commits/d16-3
[1]: https://github.com/xamarin/xamarin-android/commit/27b25ef416c5ace9b90153c192db1de6f80b1296