[Xamarin.Android.Build.Tasks] fix for empty $(AndroidNdkDirectory)
Context: https://github.com/xamarin/xamarin-android/pull/3402

When reviewing #3402, we found two things that seem wrong™:

* A case where `AndroidNdkDirectory` in the
  `<GetAdditionalResourcesFromAssemblies/>` MSBuild task, was
  `[Required]`. How did that ever work?
* In cases where `$(AndroidNdkDirectory)` is blank, it appears to get
  set to `\`. Ok that's how it worked...

The first fix was to just remove `[Required]`, I did not find any
other places this was incorrect. I also had to adjust
`<GetAdditionalResourcesFromAssemblies/>` to account for the value
being blank.

Then I revisited the pattern we have in the `_ResolveMonoAndroidSdks`
MSBuild target:

    <CreateProperty Value="$(AppSettingsDirectory)">
        <Output TaskParameter="Value" PropertyName="_AppSettingsDirectory"/>
    </CreateProperty>
    <CreateProperty Value="$(_AppSettingsDirectory)\">
        <Output TaskParameter="Value" PropertyName="_AppSettingsDirectory"
            Condition="!HasTrailingSlash('$(_AppSettingsDirectory)')" />
    </CreateProperty>

That is a lot of logic, which actually doesn't work property if the
incoming value is blank... The end result is `\`.

This would be more properly written as:

    <_AppSettingsDirectory>$([MSBuild]::EnsureTrailingSlash($(AppSettingsDirectory)))</_AppSettingsDirectory>

https://docs.microsoft.com/en-us/visualstudio/msbuild/property-functions?view=vs-2019#msbuild-property-functions

Using `$([MSBuild]::EnsureTrailingSlash())` does the following:

    If the given path doesn't have a trailing slash then add one. If the path is an empty string, does not modify it.

I have reworked the `_ResolveMonoAndroidSdks` target to use this
simpler pattern instead.

I also wrote a few tests around this target, which runs the
`_ResolveMonoAndroidSdks` target directly. Each test takes < 1 second
to complete, since it isn't actually doing a full build.