[Xamarin.Android.Build.Tests] Run tests in the right directory. (#3436)
We are currently always running the tests in the
`bin\TestRelease\temp` directory. However it seems our
build system is restoring Nuget packages into

    bin\Test$(Configuration)\temp\packages

As a result for certain tests the `%(HintPath)`'s are no
longer valid and that causes the test to fail.

This commit changes the path to make use of the
`$(Configuration)` so we end up using the correct
output path for the tests.