[Xamarin.Android.Build.Tasks] fix a case where armeabi is picked up
Fixes: https://github.com/xamarin/xamarin-android/issues/3373

A certain situation was causing `armeabi` to be treated as a valid ABI
for native libraries:

* Using a NuGet package that includes an `armeabi` .so file.
* Some directory has a name of a *valid* abi, such as
  `$(IntermediateOutputPath)` set to `obj\Debug\x86`.

`MonoAndroidHelper.GetNativeLibraryAbi` was looking at all directory
names -- and so if `x86` was in the path at all, `armeabi/libfoo.so`
will be used throughout the build.

I reworked `GetNativeLibraryAbi` so that it only looks at the topmost
directory that contains the `.so` file. I also added a test for this
scenario.