[tests] SetLastWriteTimeUtc in InvalidAndroidResource test (#3447)
Context: http://build.azdo.io/2915638

The `IncrementalBuildTest.InvalidAndroidResource()` test has been
failing randomly with:

	Resources/values/ids.xml(2): error APT0000: <item> must have a 'type' attribute.
	Resources/values/ids.xml : error APT0000: file failed to compile.

This test has a few steps:

 1. Build a project with the above error -- ignore the error
 2. Remove the offending file, build again
 3. Add the offending file back -- ignore the error

It looks like on step No. 2 that the `.csproj` file for some reason is
not having a timestamp that would trigger this target:

	Target Name=_GenerateAndroidResourceDir Project=UnnamedProject.csproj
	Skipping target "_GenerateAndroidResourceDir" because all output files are up-to-date with respect to the input files.

Normally, this target runs during this test, due to the `.csproj` file
having a newer timestamp, and this MSBuild task will delete the
`ids.xml` file:

	<RemoveUnknownFiles
	    Files="@(_AndroidResourceDest)"
	    Directory="$(MonoAndroidResDirIntermediate)"
	    RemoveDirectories="true">
	  <Output ItemName="_AndroidResourceDestRemovedFiles" TaskParameter="RemovedFiles" />
	</RemoveUnknownFiles>

This test failure is only happening randomly on macOS, which seems
like it is a bug with `Xamarin.ProjectTools` running on Mono.
For now, we can use this workaround to get our test run green again.