[ci] Remove monodroid provisionator script
We currently use this script to install the JDK and select a specific
version of Xcode, however these dependencies should no longer be needed.
We are already installing the Corretto JDK as part of xaprepare, and we
should no longer have a dependency on Xcode 9.2.

As of commit 08e0ccb4 `$(AndroidToolchainDirectory)` is set to
`$(HOME)/Library/Android` when executing on macOS Azure Pipelines hosts.
The `$(JAVA_HOME)` variable set during our core macOS build has been
updated to point to this path which will contain the Corretto JDK.