[docs] Prepare messages README for sync to docs.microsoft.com
Context: https://github.com/MicrosoftDocs/xamarin-docs/blob/live/contributing-guidelines/template.md

Add a metadata block with the title of the page, a description, and the
date of the most recent page update.

Add a first-level title heading.

Change the third-level section headings to second-level headings so they
will automatically generate an on-page table of contents when published
on docs.microsoft.com.

Adjust the section headings to use sentence casing.

Adjust "Adb" and "Aapt" to all caps "ADB" and "AAPT" to match how they
are capitalized when used as acronyms on developer.android.com.